+++ app/src/main/java/elite/intel/ai/brain/gemini/GeminiChatEndPoint.java
@@ -80,7 +80,7 @@ public class GeminiChatEndPoint extends AiEndPoint implements AIChatInterface {
             if (!rawResponse.has("candidates")) {
                 log.error("No candidates in Gemini response:\n{}", rawResponse);
                 JsonObject fallback = new JsonObject();
-                fallback.addProperty(AIConstants.PROPERTY_TEXT_TO_SPEECH_RESPONSE, "Gemeni did not return any candidates. LLM Code aborted");
+                fallback.addProperty(AIConstants.PROPERTY_TEXT_TO_SPEECH_RESPONSE, "Gemini did not return any candidates. LLM Code aborted");
                 return fallback;
             }